dd_fit_stations

Fit a generalised exponential decay model to data from one station

dd_fit_stations(city, from = TRUE, lower = 0, upper = 1, mi = FALSE,
  expmod = TRUE, osm = TRUE, plot = FALSE)

Arguments

city
City for which model is to be fitted

from
Analyse decay functions for trips from each station. If FALSE, analyse for trips to each station.

lower
Lower limit (0-1) for distance cutoff used to calculate covariances (see details)

upper
Upper limit (0-1) for distance cutoff used to calculate covariances (see details)

mi
If TRUE, fit decay functions for mutual information, otherwise fit covariances.

expmod
If TRUE, fit exponential decay models, otherwise fit power-law decays.

osm
If FALSE, use straight-line distances for distance decay parameters, otherwise street network distances.

plot
If TRUE, plot decay function

Value

A data.frame of four values: id, the ID of the station; k, the width parameter of the exponential decay; b, the value of the exponent; and ss, the standardised sum of squared residuals.

Note

Power-law fits (with expmod = FALSE) are not reliable at all. The decays really are not power-laws, so this ought not be used.