Calculate distance matrix between pairs of points using the google API

Calculate distance matrix between pairs of points using the google API

distmat_g(xy, dmat, g_units = "metric", g_mode = "bicycling")

Arguments

xy
A two-column matrix of lon-lat coordinates

dmat
If submitted, any missing values not returned in previous queries are filled.

g_units
The units of measurement requested from the google API

g_mode
The mode of transport requested from the google API

Note

the google server frequently returns no values, so this produces a matrix with lots of missing values.

Examples

# NOT RUN {
test_fn ()
# }